requirements.txt
addict
future
lmdb
setuptools>=57.0.0
numpy>=1.20.3
opencv-python-headless
Pillow>=8.3.2
pyyaml
requests
scikit-image
scikit-learn
scipy
tb-nightly
torch>=1.8.1
torchvision>=0.9
tqdm
yapf
pyiqa
einops
